apps/diffusion/requirements.txt
flask
flask_cors
bentoml>=1.2.17
pillow==10.3.0
git+https://github.com/huggingface/diffusers
wandb
bitsandbytes
safetensors
accelerate>=0.22.0
torch==2.3.1
torchvision
transformers>=4.25.1
ftfy
tensorboard
Jinja2
datasets
peft==0.7.0